Let portalLib.{getComponent,getContent} not return null
The documentation states that portalLib.getComponent() and
portalLib.getContent() should be used in parts, layouts and pages.

In these cases they would never return null. If they are misused in
webapps or error-handlers, the type system will not help prevent
null-errors, but it might be better if the function fails quickly.